Implemented converting i8 to u8

+++ src/plugins/intel_cpu/tests/functional/shared_tests_instances/snippets/convert.cpp
@@ -14,7 +14,7 @@ namespace {
 
 #if 1
 const std::vector<std::pair<std::vector<ov::element::Type>, std::vector<ov::element::Type>>> types_Convert = {
-        { { ov::element::i8 }, { ov::element::f16 } },
+        { { ov::element::i8 }, { ov::element::u8 } },
 };
 
 const std::vector<std::vector<ov::test::InputShape>> inputShapes_Convert = {
